requirements.txt
backtrader==1.9.78.123
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
mlflow==2.10.2
pandas==2.2.0
boto3==1.34.34
ib_insync==0.9.85
oandapyV20==0.7.2
pytest==8.0.0
